Cast size_t to int strategically
I think we can safely assume there are fewer than 2147483648 tests in a
suite.

I experimented with exposing size_t all the way through to clients, but
that made the testing of testresults.h itself much messier, e.g.

    Assert::That(results.NumberOfTestsRun(),
        Is().EqualTo(0));

vs

    Assert::That(results.NumberOfTestsRun(),
        Is().EqualTo(size_t(0)));

It could be argued that's a bug in snowhouse, that it requires perfect
type matching for expected/actual.

Closes #13.